added transfer and link function

+++ src/main/java/com/orkblan/FlashCash/domain/Account.java
@@ -9,7 +9,7 @@ public class Account {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long account_id;
-    private Integer solde;
+    private Double solde;
     private String iban;
     @OneToOne(cascade = CascadeType.ALL)
     @JoinColumn(name = "user_id", referencedColumnName = "user_id")
